Cerveau en compote, new beat is definetively more robust but it is not working just yet. At least, at this moment, there is a passive beat at 120 BPM!
EXPLANATION OF THE LOGIC: there are passvie beats and active beats. passive beats beat at 120 bpm all the time, and active beat depend on the Bar. the bar is an array of 8 bool that represent beats and offbeats. example: if you want to make a simple base beat, the bar should be {true, false, true, false, true, false, true, false}. That's where I am so far.